Very quick refactoring to fix warnings

+++ n-BodySim/button.h
@@ -16,6 +16,6 @@ public:
 private:
 	bool isOnTop = false;
 	Color color = { 128,128,128,255 };
-	int textSize = 15;
-	int textSeparation = 4;
+	float textSize = 15.0f;
+	float textSeparation = 4.0f;
 };
